Fixed hardware disk to double

+++ iaas-common/src/main/java/io/github/cloudiator/iaas/common/domain/NodeProperties.java
@@ -10,7 +10,7 @@ public interface NodeProperties {
 
   long memory();
 
-  Optional<Float> disk();
+  Optional<Double> disk();
 
   Optional<OperatingSystem> operatingSystem();
